+++ src/ckanapi_harvesters/harvesters/file_formats/file_format_init.py
@@ -6,11 +6,15 @@ File format keyword selection
 from ckanapi_harvesters.harvesters.file_formats.file_format_abc import FileFormatABC
 from ckanapi_harvesters.harvesters.file_formats.csv_format import CsvFileFormat
 from ckanapi_harvesters.harvesters.file_formats.shp_format import ShapeFileFormat
+from ckanapi_harvesters.harvesters.file_formats.xls_format import ExcelFileFormat
+from ckanapi_harvesters.harvesters.file_formats.json_format import JsonFileFormat
 from ckanapi_harvesters.harvesters.file_formats.user_format import UserFileFormat
 
 file_format_dict = {
     "csv": CsvFileFormat,
     "shp": ShapeFileFormat,
+    "xls": ExcelFileFormat,
+    "json": JsonFileFormat,
 }
 
 def init_file_format_datastore(format:str, options_string:str=None, aux_read_fun_name:str=None, aux_write_fun_name:str=None) -> FileFormatABC: