+++ src/ckanapi_harvesters/harvesters/file_formats/user_format.py
@@ -27,8 +27,9 @@ def write_function_example(df: Union[pd.DataFrame, List[dict]], file_path_or_buf
 
 class UserFileFormat(FileFormatABC):
     def __init__(self, options_string: str, *, df_read_fun:Callable[[Any], GeneralDataFrame] = None,
-                 df_write_fun:Callable[[GeneralDataFrame, Any], Any] = None) -> None:
-        super().__init__(options_string=options_string)
+                 df_write_fun:Callable[[GeneralDataFrame, Any], Any] = None,
+                 read_kwargs:dict=None, write_kwargs:dict=None) -> None:
+        super().__init__(options_string=options_string, read_kwargs=read_kwargs, write_kwargs=write_kwargs)
         self.df_read_fun:Union[Callable[[Any], GeneralDataFrame], None] = df_read_fun
         self.df_write_fun:Union[Callable[[pd.DataFrame, Any], pd.DataFrame], None] = df_write_fun
         self.option_append_allowed: bool = False
@@ -39,6 +40,7 @@ class UserFileFormat(FileFormatABC):
         parser = super()._setup_cli_parser(parser)
         parser.add_argument("--allow-append",
                             help="Option to signal the append mode is available for the write function", action="store_true", default=False)
+        return parser
 
     def _apply_arguments(self, args: argparse.Namespace, extra_args: list):
         super()._apply_arguments(args, extra_args)
@@ -54,21 +56,27 @@ class UserFileFormat(FileFormatABC):
             self.df_write_fun = module.function_pointer(aux_write_fun_name)
 
     # read -------------------
+    def read_by_chunks_allowed(self) -> bool:
+        return True
+
     def read_file(self, file_path: str, fields: Union[Dict[str, CkanField],None], allow_chunks:bool=True) -> Union[pd.DataFrame, ListRecords]:
         if self.df_read_fun is None:
             raise MissingIOFunctionError("Read function")
-        return self.df_read_fun(file_path, fields=fields, allow_chunks=allow_chunks and self.allow_chunks, params=self, **self.read_kwargs)
+        read_kwargs = self._get_read_kwargs(allow_chunks=allow_chunks)
+        return self.df_read_fun(file_path, fields=fields, allow_chunks=self.read_by_chunks_enabled(allow_chunks=allow_chunks), params=self, **read_kwargs)
 
     def read_buffer_full(self, buffer: io.StringIO, fields: Union[Dict[str, CkanField],None]) -> Union[pd.DataFrame, ListRecords]:
         if self.df_read_fun is None:
             raise MissingIOFunctionError("Read function")
-        return self.df_read_fun(buffer, fields=fields, allow_chunks=False, params=self, **self.read_kwargs)
+        read_kwargs = self._get_read_kwargs(allow_chunks=False)
+        return self.df_read_fun(buffer, fields=fields, allow_chunks=False, params=self, **read_kwargs)
 
     # write ------------------
     def write_file(self, df: pd.DataFrame, file_path: str, fields: Union[Dict[str, CkanField],None]) -> None:
         if self.df_write_fun is None:
             raise MissingIOFunctionError("Write function")
-        self.df_write_fun(df, file_path, append=False, fields=fields, params=self, **self.write_kwargs)
+        write_kwargs = self._get_write_kwargs()
+        self.df_write_fun(df, file_path, append=False, fields=fields, params=self, **write_kwargs)
 
     def append_allowed(self) -> bool:
         return self.option_append_allowed
@@ -77,26 +85,29 @@ class UserFileFormat(FileFormatABC):
                     fields: Union[Dict[str, CkanField], None]) -> None:
         if self.df_write_fun is None:
             raise MissingIOFunctionError("Write function")
-        self.df_write_fun(df, file_path, append=True, fields=fields, params=self, **self.write_kwargs)
+        write_kwargs = self._get_write_kwargs()
+        self.df_write_fun(df, file_path, append=True, fields=fields, params=self, **write_kwargs)
 
     def write_in_memory(self, df: pd.DataFrame, fields: Union[Dict[str, CkanField],None]) -> bytes:
         if self.df_write_fun is None:
             raise MissingIOFunctionError("Write function")
         buffer = io.StringIO()
-        self.df_write_fun(df, buffer, append=False, fields=fields, params=self, **self.write_kwargs)
+        write_kwargs = self._get_write_kwargs()
+        self.df_write_fun(df, buffer, append=False, fields=fields, params=self, **write_kwargs)
         return buffer.getvalue().encode("utf8")
 
     def append_in_memory(self, buffer: bytes, df: Union[pd.DataFrame, ListRecords], fields: Union[Dict[str, CkanField],None]) -> bytes:
         if self.df_write_fun is None:
             raise MissingIOFunctionError("Write function")
         buffer = io.StringIO(buffer.decode("utf8"))
-        self.df_write_fun(df, buffer, append=True, fields=fields, params=self, **self.write_kwargs)
+        write_kwargs = self._get_write_kwargs()
+        self.df_write_fun(df, buffer, append=True, fields=fields, params=self, **write_kwargs)
         return buffer.getvalue().encode("utf8")
 
     # misc ------------------
     def copy(self, dest=None):
         if dest is None:
-            dest = UserFileFormat(self.options_string)
+            dest = UserFileFormat(self.options_string, read_kwargs=self.read_kwargs, write_kwargs=self.write_kwargs)
         super().copy(dest=dest)
         dest.df_read_fun = self.df_read_fun
         dest.df_write_fun = self.df_write_fun