+++ CHANGELOG.md
@@ -14,10 +14,13 @@ When publishing a new release, copy the relevant section on the [Github release 
 - Support for reading CSV files by chunks using pandas.read_csv `chunksize` argument (only for DataStores). 
 This modification is compatible with multi-threading and returns a more accurate progression indicator.
 It is enabled by default. To disable it, use the `allow_chunks=False` argument or enter `--no-chunks` in the __Options__ column of the resources worksheet.
-- Excel builder workbook:
+- File formats
+  - Added CLI argument `--read-kwargs`/`--write-kwargs` in the __Options__ column to customize the arguments for the read/write functions.
+  e.g. for CSV file format, the arguments of the `pandas.read_csv` function can be changed with the following 
+  CLI argument `--read-kwargs compression=gzip,header=10`.
+  - Added support for the following file formats: Excel (xls, xlsx, xlsm, xlsb, odf, ods, odt), JSON. 
   - Support for user-defined file format loading functions with new Excel columns __Read function__ and __Write function__ per resource.
-  - Support for extra file reading options to change the arguments of the `pandas.read_csv` function: 
-  CLI argument `--read-kwargs` in the __Options__ column - e.g. `--read-kwargs compression=gzip,header=10`.
+- Excel builder workbook:
   - Extra column __Data cleaner__ per DataStore to activate a function which corrects values according to the destination type specified in the fields metadata (only for uploads). 
   A CLI option `--data-cleaner` allows setting this property with the __Options__ column.
   - CKAN upload parameterization fields (_ckan_ sheet):